fmt: Move all code into toggl2clockify subdirectory, cleanup main migrate file, shorten entrypoint file.
